+++ srcts/types/src/inputPolicies/inputRatePolicy.d.ts
@@ -1,0 +1,9 @@
+import type { AnyVoidFunction } from "../utils/extraTypes";
+import type { InputPolicy } from "./inputPolicy";
+interface InputRatePolicy<X extends AnyVoidFunction> {
+    target: InputPolicy;
+    func: X;
+    normalCall(...args: Parameters<X>): void;
+    immediateCall(...args: Parameters<X>): void;
+}
+export type { InputRatePolicy };